try errors return handler, instead of fatal errors aborting
So far running Alltoall on CRAY MPI with one rank passing nullptr causes
abort, but the SLURM job still keeps going, so it seems like the other
ranks are caught in a deadlock. It seems like this return error handler
still gives a similar result (deadlock), but the rank with error can
keep going and print out it's error code, rather than having "MPICH
ERROR [Rank 0]..." show in the log.